Add endDate check in request list, update minor codes in controllers

+++ src/main/java/org/smu/blood/controller/MyPageController.java
@@ -1,8 +1,6 @@
 package org.smu.blood.controller;
 
 import java.util.HashMap;
-import java.util.Optional;
-
 import org.smu.blood.api.JWTService;
 import org.smu.blood.database.User;
 import org.smu.blood.database.UserRepository;
@@ -50,39 +48,48 @@ public class MyPageController {
 			// token에서 사용자 id 가져오기
 			String userId = jwtService.getClaim(token).get("id").toString();
 			System.out.println("[+] userId from token: " + userId);
-			User user = repository.findById(userId).get();
-			if(user != null) { System.out.println("[+] get user info: "+ user.toString()); return user; }
-			else return null;
+
+			if(repository.findById(userId).isPresent()){
+				User user = repository.findById(userId).get();
+				System.out.println("[+] get user info: "+ user);
+				return user;
+			} else{
+				System.out.println("[-] get user info: failed");
+			}
 		}else {
 			System.out.println("[-] Invalid token");
-			return null;
 		}
+		// if no user or invalid token
+		return null;
 	}
 	// 내 정보 수정
 	@PostMapping("myPage/edit")
 	public HashMap<String,Integer> setMyData(@RequestHeader String token, @RequestBody HashMap<String,String> editInfo) {
 		System.out.println("[+] Connection from Android");
 		System.out.println("[+] token: " + token);
-		HashMap<String,Integer> result = new HashMap<String,Integer>();
+		HashMap<String,Integer> result = new HashMap<>();
 		
 		// token 유효성 검증
 		if(jwtService.checkTokenExp(token)) {
 			// 사용자 id 가져오기
 			String userId = jwtService.getClaim(token).get("id").toString();
 			System.out.println("[+] userId from token: " + userId);
-			// 사용자 정보 내 패스워드와 사용자 입력 패스워드 일치 확인 
-			if(repository.findByIdAndPassword(userId, editInfo.get("current_pw")).equals(Optional.empty())){
+
+			// 사용자 정보 내 패스워드와 사용자 입력 패스워드 일치하지 않는 경우
+			if(!repository.findByIdAndPassword(userId, editInfo.get("current_pw")).isPresent()){
 				System.out.println("Incorrect password");
 				result.put("wrong_pw", 1);
-			}else {
+			}else { // 사용자 정보 내 패스워드와 사용자 입력 패스워드 일치하는 경우
 				User user = repository.findByIdAndPassword(userId, editInfo.get("current_pw")).get();
-				System.out.println("Current User: "+user.toString());
+				System.out.println("Current User: "+user);
+
 				// 변경 비밀번호 값 있는 경우
 				if(editInfo.containsKey("new_pw")) user.setPassword(editInfo.get("new_pw"));
 				// 변경 닉네임 값 있는 경우
 				if(editInfo.containsKey("new_nickname")) user.setNickname(editInfo.get("new_nickname"));
-				System.out.println("Update User: "+user.toString());
+
 				// 새로운 패스워드 또는 닉네임으로 사용자 정보 업데이트
+				System.out.println("Update User: "+user);
 				repository.save(user);
 			}
 		}else {
@@ -101,12 +108,17 @@ public class MyPageController {
 			// 사용자 id 가져오기
 			String userId = jwtService.getClaim(token).get("id").toString();
 			System.out.println("[+] userId from token: " + userId);
-			repository.delete(repository.findById(userId).get());
-			System.out.println("[+] User Deletion Success");
-			return true;
+
+			if(repository.findById(userId).isPresent()){
+				repository.delete(repository.findById(userId).get());
+
+				System.out.println("[+] User Deletion Success");
+				return true;
+			}else System.out.println("[-] no user info. deletion failed.");
 		}else {
 			System.out.println("[-] Invalid token");
-			return false;
 		}
+		// invalid token or no user info to delete
+		return false;
 	}
 }
